major refactoring for logging
Now using standard logging module and a single log file. A Single log file has
great advantages when having troubles with race conditions. Also the logging
module makes life much easier.

TODO: We do not use the file name in robot.cfg because we do not want to parse
the cfg files on import time. Don't know how to solve this in a nice way... :(
(A funny but overcomplicated way would be to use a BufferHandler and to buffer
until config is parsed.)